Bug 1407911 - update to v0.2
Update to 0.2 so we can do some QA on Beta 57.

The v0.2 will set stylo blocklist to "arewestyloyet.rs", which is the website
designed for verifing stylo blocklist's functionality.

I'm keeping both version (v0.1 and v0.2) for now, as a backup. Will squash the
commit to only one version once we need to publish v1.0.